Move static MF test to Static's test project and address some comments
