+++ test/Microsoft.ML.Tests/TrainerEstimators/MatrixFactorizationTests.cs
@@ -2,12 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-
 using Microsoft.ML.Runtime.Api;
 using Microsoft.ML.Runtime.Data;
 using Microsoft.ML.Runtime.RunTests;
-using Microsoft.ML.Runtime.Recommender;
-using Microsoft.ML.StaticPipe;
 using Microsoft.ML.Trainers;
 using System;
 using System.Collections.Generic;
@@ -50,125 +47,74 @@ namespace Microsoft.ML.Tests.TrainerEstimators
         [Fact]
         public void MatrixFactorizationSimpleTrainAndPredict()
         {
-            using (var env = new LocalEnvironment(seed: 1, conc: 1))
-            {
-                // Specific column names of the considered data set
-                string labelColumnName = "Label";
-                string userColumnName = "User";
-                string itemColumnName = "Item";
-                string scoreColumnName = "Score";
-
-                // Create reader for both of training and test data sets
-                var reader = new TextLoader(env, GetLoaderArgs(labelColumnName, userColumnName, itemColumnName));
-
-                // Read training data as an IDataView object
-                var data = reader.Read(new MultiFileSource(GetDataPath(TestDatasets.trivialMatrixFactorization.trainFilename)));
-
-                // Create a pipeline with a single operator.
-                var pipeline = new MatrixFactorizationTrainer(env, labelColumnName, userColumnName, itemColumnName, 
-                    advancedSettings:s=>
-                    {
-                        s.NumIterations = 3;
-                        s.NumThreads = 1; // To eliminate randomness, # of threads must be 1.
-                        s.K = 7;
-                    });
-
-                // Train a matrix factorization model.
-                var model = pipeline.Fit(data);
-
-                // Read the test data set as an IDataView
-                var testData = reader.Read(new MultiFileSource(GetDataPath(TestDatasets.trivialMatrixFactorization.testFilename)));
-
-                // Apply the trained model to the test set
-                var prediction = model.Transform(testData);
-
-                // Get output schema and check its column names
-                var outputSchema = model.GetOutputSchema(data.Schema);
-                var expectedOutputNames = new string[] { labelColumnName, userColumnName, itemColumnName, scoreColumnName };
-                foreach (var (i, col) in outputSchema.GetColumns())
-                    Assert.True(col.Name == expectedOutputNames[i]);
-
-                // Retrieve label column's index from the test IDataView
-                testData.Schema.TryGetColumnIndex(labelColumnName, out int labelColumnId);
-
-                // Retrieve score column's index from the IDataView produced by the trained model
-                prediction.Schema.TryGetColumnIndex(scoreColumnName, out int scoreColumnId);
-
-                // Compute prediction errors
-                var mlContext = new MLContext();
-                var metrices = mlContext.Regression.Evaluate(prediction, label: labelColumnName, score: scoreColumnName);
-
-                // Determine if the selected metric is reasonable for different platforms
-                double tolerance = Math.Pow(10, -7);
-                if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-                {
-                    // Unix case
-                    var expectedUnixL2Error = 0.616821448679879; // Unix baseline
-                    Assert.InRange(metrices.L2, expectedUnixL2Error - tolerance, expectedUnixL2Error + tolerance);
-                }
-                else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
-
-                {
-                    // Mac case
-                    var expectedMacL2Error = 0.61192207960271; // Mac baseline
-                    Assert.InRange(metrices.L2, expectedMacL2Error - 5e-3, expectedMacL2Error + 5e-3); // 1e-7 is too small for Mac so we try 1e-5
-                }
-                else if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-                {
-                    // Windows case
-                    var expectedWindowsL2Error = 0.61528733643754685; // Windows baseline
-                    Assert.InRange(metrices.L2, expectedWindowsL2Error - tolerance, expectedWindowsL2Error + tolerance);
-                }
-            }
-        }
-
-        [Fact]
-        public void MatrixFactorizationStatic()
-        {
-            // Create a new context for ML.NET operations. It can be used for exception tracking and logging, 
-            // as a catalog of available operations and as the source of randomness.
             var mlContext = new MLContext(seed: 1, conc: 1);
 
-            // Specify where to find data file
-            var dataPath = GetDataPath(TestDatasets.trivialMatrixFactorization.trainFilename);
-            var dataSource = new MultiFileSource(dataPath);
+            // Specific column names of the considered data set
+            string labelColumnName = "Label";
+            string userColumnName = "User";
+            string itemColumnName = "Item";
+            string scoreColumnName = "Score";
 
-            // Read data file. The file contains 3 columns, label (float value), matrixColumnIndex (unsigned integer key), and matrixRowIndex (unsigned integer key).
-            // More specifically, LoadUInt32Key(1, 0, 19) means that the matrixColumnIndex column is read from the 2nd (indexed by 1) column in the data file and as
-            // a key type (stored as 32-bit unsigned integer) ranged from 0 to 19 (aka the training matrix has 20 columns).
-            var reader = mlContext.Data.TextReader(ctx => (label: ctx.LoadFloat(0), matrixColumnIndex: ctx.LoadUInt32Key(1, 0, 19), matrixRowIndex: ctx.LoadUInt32Key(2, 0, 39)));
+            // Create reader for both of training and test data sets
+            var reader = new TextLoader(mlContext, GetLoaderArgs(labelColumnName, userColumnName, itemColumnName));
 
-            // The parameter that will be into the onFit method below. The obtained predictor will be assigned to this variable
-            // so that we will be able to touch it.
-            MatrixFactorizationPredictor pred = null;
+            // Read training data as an IDataView object
+            var data = reader.Read(new MultiFileSource(GetDataPath(TestDatasets.trivialMatrixFactorization.trainFilename)));
 
-            // Create a statically-typed matrix factorization estimator. The MatrixFactorization's input and output defined in MatrixFactorizationStatic
-            // tell what (aks a Scalar<float>) is expected. Notice that only one thread is used for deterministic outcome.
-            var matrixFactorizationEstimator = reader.MakeNewEstimator()
-                .Append(r => (r.label, score: mlContext.Regression.Trainers.MatrixFactorization(r.label, r.matrixRowIndex, r.matrixColumnIndex, onFit: p => pred = p,
-                advancedSettings: args => { args.NumThreads = 1; })));
+            // Create a pipeline with a single operator.
+            var pipeline = new MatrixFactorizationTrainer(mlContext, labelColumnName, userColumnName, itemColumnName, 
+                advancedSettings:s=>
+                {
+                    s.NumIterations = 3;
+                    s.NumThreads = 1; // To eliminate randomness, # of threads must be 1.
+                    s.K = 7;
+                });
 
-            // Create a pipeline from the reader (the 1st step) and the matrix factorization estimator (the 2nd step).
-            var pipe = reader.Append(matrixFactorizationEstimator);
+            // Train a matrix factorization model.
+            var model = pipeline.Fit(data);
 
-            // pred will be assigned by the onFit method once the training process is finished, so pred must be null before training.
-            Assert.Null(pred);
+            // Read the test data set as an IDataView
+            var testData = reader.Read(new MultiFileSource(GetDataPath(TestDatasets.trivialMatrixFactorization.testFilename)));
 
-            // Train the pipeline on the given data file. Steps in the pipeline are sequentially fitted (by calling their Fit function).
-            var model = pipe.Fit(dataSource);
+            // Apply the trained model to the test set
+            var prediction = model.Transform(testData);
 
-            // pred got assigned so that one can inspect the predictor trained in pipeline.
-            Assert.NotNull(pred);
+            // Get output schema and check its column names
+            var outputSchema = model.GetOutputSchema(data.Schema);
+            var expectedOutputNames = new string[] { labelColumnName, userColumnName, itemColumnName, scoreColumnName };
+            foreach (var (i, col) in outputSchema.GetColumns())
+                Assert.True(col.Name == expectedOutputNames[i]);
 
-            // Feed the data file into the trained pipeline. The data would be loaded by TextLoader (the 1st step) and then the output of the
-            // TextLoader would be fed into MatrixFactorizationEstimator.
-            var estimatedData = model.Read(dataSource);
+            // Retrieve label column's index from the test IDataView
+            testData.Schema.TryGetColumnIndex(labelColumnName, out int labelColumnId);
 
-            // After the training process, the metrics for regression problems can be computed.
-            var metrics = mlContext.Regression.Evaluate(estimatedData, r => r.label, r => r.score);
+            // Retrieve score column's index from the IDataView produced by the trained model
+            prediction.Schema.TryGetColumnIndex(scoreColumnName, out int scoreColumnId);
 
-            // Naive test. Just make sure the pipeline runs.
-            Assert.True(metrics.L2 > 0);
+            // Compute prediction errors
+            var metrices = mlContext.Regression.Evaluate(prediction, label: labelColumnName, score: scoreColumnName);
+
+            // Determine if the selected metric is reasonable for different platforms
+            double tolerance = Math.Pow(10, -7);
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+            {
+                // Linux case
+                var expectedUnixL2Error = 0.616821448679879; // Linux baseline
+                Assert.InRange(metrices.L2, expectedUnixL2Error - tolerance, expectedUnixL2Error + tolerance);
+            }
+            else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
+
+            {
+                // Mac case
+                var expectedMacL2Error = 0.61192207960271; // Mac baseline
+                Assert.InRange(metrices.L2, expectedMacL2Error - 5e-3, expectedMacL2Error + 5e-3); // 1e-7 is too small for Mac so we try 1e-5
+            }
+            else if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+            {
+                // Windows case
+                var expectedWindowsL2Error = 0.61528733643754685; // Windows baseline
+                Assert.InRange(metrices.L2, expectedWindowsL2Error - tolerance, expectedWindowsL2Error + tolerance);
+            }
         }
 
         private TextLoader.Arguments GetLoaderArgs(string labelColumnName, string matrixColumnIndexColumnName, string matrixRowIndexColumnName)
